requirements.txt
streamlit
python-chess
cairosvg
pillow